♻️ refactor: Relocate GraphRAG methods and extract utility function
Move async GraphRAG search methods from DVS class to Graph API class for better code organization. Extract cosine similarity calculation into dedicated utility module. Update example to use new method locations.

This refactoring improves separation of concerns by centralizing graph-related functionality in the appropriate module.
